Tarefas feitas
The most significant changes were made to the `agenda-profissional.component.ts`, `agenda-profissional.component.css`, and `agenda-profissional.component.html` files. These changes include updates to the `changeAppointmentScheduled` and `cancelAppointment` functions, the addition of new functions, and the inclusion of new CSS styles and HTML elements.

In the `agenda-profissional.component.ts` file, the `changeAppointmentScheduled` and `cancelAppointment` functions were updated, and new functions `openPopup`, `closePopup`, and `stopPropagation` were added.

The `agenda-profissional.component.css` file was updated with new CSS styles for several classes including overlay, close-button, close-image, overlay-container, speciality-form-box, speciality-name, speciality-select-div, form-buttons-row, btns, btns-blue, and btns-red.

The `agenda-profissional.component.html` file was updated to include changes in the button class and id, and to include new div elements for the overlay, remove-appointment-container, and accept-appointment-container.

Other changes include updates to the `agenda-paciente.component.ts`, `consulta-paciente.component.html`, `consulta-profissional.component.html`, `home-page.component.html`, and `marcacoes.component.ts` files.

In the `agenda-paciente.component.ts` file, a new line of code was added that sets the `appointmentsFiltered` variable to `this.appointments` and calls the `updateDisplayedAppointments` function.

The `consulta-paciente.component.html` and `consulta-profissional.component.html` files were updated to change the text on the buttons in the remove-appointment-container div.

The `home-page.component.html` file was updated to include a routerLink in the anchor tag within the li element.

The `marcacoes.component.ts` file was updated to change the value of `itemsPerPage` from 10 to 15.

Unfortunately, the document does not provide any specific changes made in the 20240418170336_LusoHealth.Designer.cs, 20240418170336_LusoHealth.cs, LusoHealth.cs, and 20240424204937_LusoHealth.Designer.cs files. Therefore, a summary of the changes cannot be provided for these files.